// src/l1/2024/types.ts
export const WK_BERUF_CODES = [
	"A",
	"B",
	"F",
	"J",
	"M",
	"FO",
	"FM",
	"HA",
	"HE",
	"V",
	"P",
	"E",
] as const;